DO_CONTAINER_DROPS only affects containers
The code for minecart drops also uses ItemScatter, preventing them from appearing if it is completely cancelled. This behavior is not expected, since the gamerule doEntityDrops exists for this purpose, and it makes the development of item transport systems more difficult, since in 1.19 it is possible to break the minecart to extract all items faster.