In the push to update the mod to 1.20.4, the gradle.properties, build.gradle, and fabric.mod.json were changed to account for any versioning.
Then the command "gradlew migrateMappings --mappings "1.20.4+build.3" was run to fix mixin errors. After that, the Gradle was updated from 8.1.1 to 8.5 due to a need by 1.20.4 Fabric.

Finally, the migrateMappings command made a new src folder and copied everything over with the fixed mixins. As such I've named the old src file "oldSrc" until we decide to delete it.